nutrition-callbot/brain/requirements.txt
# ── LLM ──
google-genai>=1.0.0
openai>=1.0.0
vllm>=0.9.0

# ── Embedding (E5-large-instruct) ──
sentence-transformers>=3.0.0
torch>=2.0.0

# ── Vector DB ──
qdrant-client>=1.7.0

# ── API Server ──
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0

# ── Compatibility ──
protobuf>=5.26.1,<6.0

# ── Utilities ──
python-dotenv>=1.0.0
